Add article opening
feat: selecting an article will open a browser with said article
This one was complicated, as I had to add a null control at Export and Feed selection because the ListBox "selected" a null value when reloading to display the new available Feeds. Will have to do for the Articles too but out of scope here.
* Also add a function to select the first article in a dict, will have to expand and propagate